+++ libraries/AP_Logger/AP_Logger_MAVLink.h
@@ -54,7 +54,7 @@ public:
 
     // high level interface
     uint16_t find_last_log(void) override { return 0; }
-    void get_log_boundaries(uint16_t log_num, uint16_t & start_page, uint16_t & end_page) override {}
+    void get_log_boundaries(uint16_t log_num, uint32_t & start_page, uint32_t & end_page) override {}
     void get_log_info(uint16_t log_num, uint32_t &size, uint32_t &time_utc) override {}
     int16_t get_log_data(uint16_t log_num, uint16_t page, uint32_t offset, uint16_t len, uint8_t *data) override { return 0; }
     uint16_t get_num_logs(void) override { return 0; }
